test: controller 통합테스트를 위한 추상 클래스 추가
