+++ src/test/kotlin/com/yapp/itemfinder/ControllerIntegrationTest.kt
@@ -1,0 +1,33 @@
+package com.yapp.itemfinder
+
+import com.fasterxml.jackson.databind.ObjectMapper
+import org.junit.jupiter.api.BeforeEach
+import org.springframework.beans.factory.annotation.Autowired
+import org.springframework.boot.test.context.SpringBootTest
+import org.springframework.test.context.ActiveProfiles
+import org.springframework.test.web.servlet.MockMvc
+import org.springframework.test.web.servlet.result.MockMvcResultHandlers
+import org.springframework.test.web.servlet.setup.DefaultMockMvcBuilder
+import org.springframework.test.web.servlet.setup.MockMvcBuilders
+import org.springframework.transaction.annotation.Transactional
+import org.springframework.web.context.WebApplicationContext
+import org.springframework.web.filter.CharacterEncodingFilter
+
+@SpringBootTest
+@Transactional
+abstract class ControllerIntegrationTest {
+    @Autowired
+    lateinit var objectMapper: ObjectMapper
+
+    lateinit var mockMvc: MockMvc
+
+    @BeforeEach
+    internal fun setUp(
+        webApplicationContext: WebApplicationContext
+    ) {
+        mockMvc = MockMvcBuilders.webAppContextSetup(webApplicationContext)
+            .addFilter<DefaultMockMvcBuilder>(CharacterEncodingFilter("UTF-8", true))
+            .alwaysDo<DefaultMockMvcBuilder>(MockMvcResultHandlers.print())
+            .build()
+    }
+}
