+++ src/test/resources/application.yml
@@ -11,7 +11,3 @@ spring:
     properties:
       hibernate:
         format_sql: true
-
-logging:
-  level:
-    org.hibernate.SQL: debug
